Add in scheduling of tasks for services
Starts new tasks on available instances when there are not enough for a
service. Does not yet kill tasks when there are too many. Additionally,
should sync task defs instead of assuming that they are available in
order to handle update services or services persisted externally that
must be remembered on app start.